NOTIF-715 Change query for orgId usage
