+++ engine/src/main/java/com/redhat/cloud/notifications/recipients/itservice/pojo/request/ITUserRequest.java
@@ -17,10 +17,8 @@ public class ITUserRequest {
 
         if (useOrgId) {
             allOf.ebsAccountNumber = null;
-            allOf.accountId = orgId;
         } else {
             allOf.ebsAccountNumber = accountId;
-            allOf.accountId = null;
         }
 
         if (adminsOnly) {
@@ -32,6 +30,12 @@ public class ITUserRequest {
         by.allOf = allOf;
 
         this.by = by;
+        if (useOrgId) {
+            this.by.accountId = orgId;
+        } else {
+            this.by.accountId = null;
+        }
+
         this.by.withPaging = new WithPaging();
         this.by.withPaging.firstResultIndex = firstResult;
         this.by.withPaging.maxResults = maxResults;
