+++ engine/src/test/java/com/redhat/cloud/notifications/recipients/itservice/pojo/request/AllOfTest.java
@@ -24,53 +24,4 @@ class AllOfTest {
 
         assertEquals("{\"ebsAccountNumber\":\"someEbsAccountNumber\",\"status\":null,\"permissionCode\":null}", result);
     }
-
-    @Test
-    void shouldContainOrgId() throws JsonProcessingException {
-        testee.accountId = "someOrgId";
-
-        String result = new ObjectMapper().writeValueAsString(testee);
-
-        assertEquals("{\"accountId\":\"someOrgId\",\"status\":null,\"permissionCode\":null}", result);
-    }
-
-    @Test
-    void shouldNotContainEbsAccountNumberWhenItIsEmpty() throws JsonProcessingException {
-        testee.accountId = "someOrgId";
-        testee.ebsAccountNumber = "";
-
-        String result = new ObjectMapper().writeValueAsString(testee);
-
-        assertEquals("{\"accountId\":\"someOrgId\",\"status\":null,\"permissionCode\":null}", result);
-    }
-
-    @Test
-    void shouldNotContainOrgIdWhenItIsEmpty() throws JsonProcessingException {
-        testee.accountId = "";
-        testee.ebsAccountNumber = "someEbsAccountNumber";
-
-        String result = new ObjectMapper().writeValueAsString(testee);
-
-        assertEquals("{\"ebsAccountNumber\":\"someEbsAccountNumber\",\"status\":null,\"permissionCode\":null}", result);
-    }
-
-    @Test
-    void shouldNotContainEbsAccountNumberWhenItIsNull() throws JsonProcessingException {
-        testee.accountId = "someOrgId";
-        testee.ebsAccountNumber = null;
-
-        String result = new ObjectMapper().writeValueAsString(testee);
-
-        assertEquals("{\"accountId\":\"someOrgId\",\"status\":null,\"permissionCode\":null}", result);
-    }
-
-    @Test
-    void shouldNotContainOrgIdWhenItIsNull() throws JsonProcessingException {
-        testee.accountId = null;
-        testee.ebsAccountNumber = "someEbsAccountNumber";
-
-        String result = new ObjectMapper().writeValueAsString(testee);
-
-        assertEquals("{\"ebsAccountNumber\":\"someEbsAccountNumber\",\"status\":null,\"permissionCode\":null}", result);
-    }
 }
